+++ CHANGES.md
@@ -4,6 +4,10 @@ Zing Changelog
 v0.5.2 (in development)
 -----------------------
 
+* Worked aaround an issue in older versions of Android+Webkit, where timezones
+  would be reported as abbreviations, and not as IANA compliant names. This
+  broke client scripts.
+
 
 v0.5.1 (2017-05-29)
 -------------------